[cpp] Split observation into multiple files
